Update README.md
This was a collaborative project executed by Gabija Safranavicius and Alvin Ghouas. It is important to note that this project is intended to function as an exploratory analysis. The suggested sociological events relayed in the analysis are hypothetical, and must be further researched before making any confident claims as to their definite influence on rap and hip-hop lyrics.